F this hubconf bs

+++ train.py
@@ -564,7 +564,8 @@ def parse_args(input_args=None):
     return args
 
 if __name__ == "__main__":
-    os.symlink('hubconf.py', '/root/.cache/torch/hub/facebookresearch_dinov2_main/hubconf.py')
+    if not os.path.exists('/root/.cache/torch/hub/facebookresearch_dinov2_main/hubconf.py'):
+        os.symlink('hubconf.py', '/root/.cache/torch/hub/facebookresearch_dinov2_main/hubconf.py')
     args = parse_args()
     print("The args are: ", args)
     exp_name = create_experiment_name(args)
